refactor: remove download widget, download directly instead
