+++ lib/boorus/danbooru/pages/widgets/danbooru_infinite_post_list.dart
@@ -142,75 +142,69 @@ class _DanbooruInfinitePostListState
               ),
               child: child,
             ),
-            child: DownloadProviderWidget(
-              builder: (context, download) => ConditionalParentWidget(
-                condition: canHandleLongPress,
-                conditionalBuilder: (child) => GestureDetector(
-                  onLongPress: () {
-                    if (postGesturesHandler != null) {
-                      postGesturesHandler(
-                        ref,
-                        ref.watchConfig.postGestures?.preview?.longPress,
-                        post,
-                        download,
-                      );
-                    }
-                  },
-                  child: child,
-                ),
-                child: ExplicitContentBlockOverlay(
-                  block: settings.blurExplicitMedia && post.isExplicit,
-                  width: width ?? 100,
-                  height: height ?? 100,
-                  childBuilder: (block) => DanbooruImageGridItem(
-                    ignoreBanOverlay: block,
-                    post: post,
-                    hideOverlay: multiSelect,
-                    autoScrollOptions: AutoScrollOptions(
-                      controller: _autoScrollController,
-                      index: index,
-                    ),
-                    onTap: !multiSelect
-                        ? () {
-                            if (booruBuilder?.canHandlePostGesture(
-                                        GestureType.tap,
-                                        booruConfig.postGestures?.preview) ==
-                                    true &&
-                                postGesturesHandler != null) {
-                              postGesturesHandler(
-                                ref,
-                                ref.watchConfig.postGestures?.preview?.tap,
-                                post,
-                                download,
-                              );
-                            } else {
-                              goToPostDetailsPage(
-                                context: context,
-                                posts: items,
-                                initialIndex: index,
-                                scrollController: _autoScrollController,
-                              );
-                            }
+            child: ConditionalParentWidget(
+              condition: canHandleLongPress,
+              conditionalBuilder: (child) => GestureDetector(
+                onLongPress: () {
+                  if (postGesturesHandler != null) {
+                    postGesturesHandler(
+                      ref,
+                      ref.watchConfig.postGestures?.preview?.longPress,
+                      post,
+                    );
+                  }
+                },
+                child: child,
+              ),
+              child: ExplicitContentBlockOverlay(
+                block: settings.blurExplicitMedia && post.isExplicit,
+                width: width ?? 100,
+                height: height ?? 100,
+                childBuilder: (block) => DanbooruImageGridItem(
+                  ignoreBanOverlay: block,
+                  post: post,
+                  hideOverlay: multiSelect,
+                  autoScrollOptions: AutoScrollOptions(
+                    controller: _autoScrollController,
+                    index: index,
+                  ),
+                  onTap: !multiSelect
+                      ? () {
+                          if (booruBuilder?.canHandlePostGesture(
+                                      GestureType.tap,
+                                      booruConfig.postGestures?.preview) ==
+                                  true &&
+                              postGesturesHandler != null) {
+                            postGesturesHandler(
+                              ref,
+                              ref.watchConfig.postGestures?.preview?.tap,
+                              post,
+                            );
+                          } else {
+                            goToPostDetailsPage(
+                              context: context,
+                              posts: items,
+                              initialIndex: index,
+                              scrollController: _autoScrollController,
+                            );
                           }
-                        : null,
-                    enableFav:
-                        !multiSelect && booruConfig.hasLoginDetails() && !block,
-                    image: BooruImage(
-                      aspectRatio: post.isBanned ? 0.8 : post.aspectRatio,
-                      imageUrl:
-                          block ? '' : post.thumbnailFromSettings(settings),
-                      borderRadius: BorderRadius.circular(
-                        settings.imageBorderRadius,
-                      ),
-                      forceFill:
-                          settings.imageListType == ImageListType.standard,
-                      placeholderUrl: post.thumbnailImageUrl,
-                      width: width,
-                      height: height,
-                      cacheHeight: cacheHeight,
-                      cacheWidth: cacheWidth,
-                      // null, // Will cause error sometimes, disabled for now
+                        }
+                      : null,
+                  enableFav:
+                      !multiSelect && booruConfig.hasLoginDetails() && !block,
+                  image: BooruImage(
+                    aspectRatio: post.isBanned ? 0.8 : post.aspectRatio,
+                    imageUrl: block ? '' : post.thumbnailFromSettings(settings),
+                    borderRadius: BorderRadius.circular(
+                      settings.imageBorderRadius,
                     ),
+                    forceFill: settings.imageListType == ImageListType.standard,
+                    placeholderUrl: post.thumbnailImageUrl,
+                    width: width,
+                    height: height,
+                    cacheHeight: cacheHeight,
+                    cacheWidth: cacheWidth,
+                    // null, // Will cause error sometimes, disabled for now
                   ),
                 ),
               ),
@@ -252,21 +246,19 @@ class FavoriteGroupMultiSelectionActions extends ConsumerWidget {
     return ButtonBar(
       alignment: MainAxisAlignment.center,
       children: [
-        DownloadProviderWidget(
-          builder: (context, download) => IconButton(
-            onPressed: selectedPosts.isNotEmpty
-                ? () {
-                    showDownloadStartToast(context);
-                    // ignore: prefer_foreach
-                    for (final p in selectedPosts) {
-                      download(p);
-                    }
-
-                    endMultiSelect();
+        IconButton(
+          onPressed: selectedPosts.isNotEmpty
+              ? () {
+                  showDownloadStartToast(context);
+                  // ignore: prefer_foreach
+                  for (final p in selectedPosts) {
+                    ref.download(p);
                   }
-                : null,
-            icon: const Icon(Symbols.download),
-          ),
+
+                  endMultiSelect();
+                }
+              : null,
+          icon: const Icon(Symbols.download),
         ),
         if (config.hasLoginDetails())
           IconButton(
