+++ lib/boorus/danbooru/pages/widgets/danbooru_multi_selection_actions.dart
@@ -11,7 +11,6 @@ import 'package:boorusama/boorus/danbooru/router.dart';
 import 'package:boorusama/core/feats/boorus/boorus.dart';
 import 'package:boorusama/core/feats/downloads/downloads.dart';
 import 'package:boorusama/core/pages/bookmarks/add_bookmarks_button.dart';
-import 'package:boorusama/core/widgets/widgets.dart';
 
 class DanbooruMultiSelectionActions extends ConsumerWidget {
   const DanbooruMultiSelectionActions({
@@ -30,21 +29,19 @@ class DanbooruMultiSelectionActions extends ConsumerWidget {
     return ButtonBar(
       alignment: MainAxisAlignment.center,
       children: [
-        DownloadProviderWidget(
-          builder: (context, download) => IconButton(
-            onPressed: selectedPosts.isNotEmpty
-                ? () {
-                    showDownloadStartToast(context);
-                    // ignore: prefer_foreach
-                    for (final p in selectedPosts) {
-                      download(p);
-                    }
-
-                    endMultiSelect();
+        IconButton(
+          onPressed: selectedPosts.isNotEmpty
+              ? () {
+                  showDownloadStartToast(context);
+                  // ignore: prefer_foreach
+                  for (final p in selectedPosts) {
+                    ref.download(p);
                   }
-                : null,
-            icon: const Icon(Symbols.download),
-          ),
+
+                  endMultiSelect();
+                }
+              : null,
+          icon: const Icon(Symbols.download),
         ),
         AddBookmarksButton(
           posts: selectedPosts,
