+++ lib/core/feats/downloads/download_provider.dart
@@ -2,6 +2,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 // Project imports:
+import 'package:boorusama/boorus/booru_builder.dart';
 import 'package:boorusama/boorus/providers.dart';
 import 'package:boorusama/core/feats/boorus/boorus.dart';
 import 'package:boorusama/core/feats/downloads/downloads.dart';
@@ -9,6 +10,9 @@ import 'package:boorusama/core/feats/posts/posts.dart';
 import 'package:boorusama/core/feats/settings/settings.dart';
 import 'package:boorusama/foundation/networking/networking.dart';
 import 'package:boorusama/foundation/path.dart';
+import 'package:boorusama/foundation/permissions.dart';
+import 'package:boorusama/foundation/platform.dart';
+import 'package:boorusama/widgets/widgets.dart';
 
 final downloadNotificationProvider = Provider<DownloadNotifications>((ref) {
   throw UnimplementedError();
@@ -59,5 +63,73 @@ String sanitizedUrl(String url) {
     return url.replaceFirst(ext, trimmedExt);
   } else {
     return url;
+  }
+}
+
+extension PostDownloadX on WidgetRef {
+  Future<void> download(Post post) async {
+    final perm = await read(deviceStoragePermissionProvider.future);
+    final settings = read(settingsProvider);
+
+    await _download(
+      this,
+      post,
+      permission: isAndroid() || isIOS() ? perm.storagePermission : null,
+      settings: settings,
+    );
+  }
+}
+
+Future<void> _download(
+  WidgetRef ref,
+  Post downloadable, {
+  PermissionStatus? permission,
+  required Settings settings,
+}) async {
+  final booruConfig = ref.readConfig;
+  final service = ref.read(downloadServiceProvider(booruConfig));
+  final fileNameBuilder =
+      ref.readBooruBuilder(booruConfig)?.downloadFilenameBuilder;
+  final downloadUrl = getDownloadFileUrl(downloadable, settings);
+
+  final logger = ref.read(loggerProvider);
+
+  if (fileNameBuilder == null) {
+    logger.logE('Single Download', 'No file name builder found, aborting...');
+    showErrorToast('Download aborted, cannot create file name');
+    return;
+  }
+
+  Future<void> download() async => service
+      .downloadWithSettings(
+        settings,
+        url: downloadUrl,
+        fileNameBuilder: () => fileNameBuilder.generate(
+          settings,
+          booruConfig,
+          downloadable,
+        ),
+      )
+      .run();
+
+  // Platform doesn't require permissions, just download it right away
+  if (permission == null) {
+    download();
+  }
+
+  if (permission == PermissionStatus.granted) {
+    download();
+  } else {
+    logger.logI('Single Download', 'Permission not granted, requesting...');
+    ref.read(deviceStoragePermissionProvider.notifier).requestPermission(
+      onDone: (isGranted) {
+        if (isGranted) {
+          download();
+        } else {
+          logger.logI('Single Download',
+              'Storage permission request denied, aborting...');
+        }
+      },
+    );
   }
 }
