+++ lib/core/scaffolds/infinite_post_list_scaffold.dart
@@ -170,96 +170,91 @@ class _InfinitePostListScaffoldState<T extends Post>
                     ),
               child: child,
             ),
-            child: DownloadProviderWidget(
-              builder: (context, download) => ConditionalParentWidget(
-                condition: canHandleLongPress,
-                conditionalBuilder: (child) => GestureDetector(
-                  onLongPress: () {
-                    if (postGesturesHandler != null) {
-                      postGesturesHandler(
-                        ref,
-                        ref.watchConfig.postGestures?.preview?.longPress,
-                        post,
-                        download,
-                      );
+            child: ConditionalParentWidget(
+              condition: canHandleLongPress,
+              conditionalBuilder: (child) => GestureDetector(
+                onLongPress: () {
+                  if (postGesturesHandler != null) {
+                    postGesturesHandler(
+                      ref,
+                      ref.watchConfig.postGestures?.preview?.longPress,
+                      post,
+                    );
+                  }
+                },
+                child: child,
+              ),
+              child: ExplicitContentBlockOverlay(
+                width: width ?? 100,
+                height: height ?? 100,
+                block: settings.blurExplicitMedia && post.isExplicit,
+                childBuilder: (block) => ImageGridItem(
+                  isGif: post.isGif,
+                  isAI: post.isAI,
+                  hideOverlay: multiSelect,
+                  onTap: !multiSelect
+                      ? () {
+                          if (booruBuilder?.canHandlePostGesture(
+                                      GestureType.tap,
+                                      config.postGestures?.preview) ==
+                                  true &&
+                              postGesturesHandler != null) {
+                            postGesturesHandler(
+                              ref,
+                              ref.watchConfig.postGestures?.preview?.tap,
+                              post,
+                            );
+                          } else {
+                            goToPostDetailsPage(
+                              context: context,
+                              posts: items,
+                              initialIndex: index,
+                              scrollController: _autoScrollController,
+                            );
+                          }
+                        }
+                      : null,
+                  isFaved: ref.watch(favoriteProvider(post.id)),
+                  enableFav: !multiSelect && canFavorite && !block,
+                  quickActionButtonBuilder:
+                      defaultImagePreviewButtonBuilder(ref, post),
+                  onFavToggle: (isFaved) async {
+                    if (isFaved) {
+                      if (favoriteAdder == null) return;
+                      await favoriteAdder(post.id, ref);
+                    } else {
+                      if (favoriteRemover == null) return;
+                      await favoriteRemover(post.id, ref);
                     }
                   },
-                  child: child,
-                ),
-                child: ExplicitContentBlockOverlay(
-                  width: width ?? 100,
-                  height: height ?? 100,
-                  block: settings.blurExplicitMedia && post.isExplicit,
-                  childBuilder: (block) => ImageGridItem(
-                    isGif: post.isGif,
-                    isAI: post.isAI,
-                    hideOverlay: multiSelect,
-                    onTap: !multiSelect
-                        ? () {
-                            if (booruBuilder?.canHandlePostGesture(
-                                        GestureType.tap,
-                                        config.postGestures?.preview) ==
-                                    true &&
-                                postGesturesHandler != null) {
-                              postGesturesHandler(
-                                ref,
-                                ref.watchConfig.postGestures?.preview?.tap,
-                                post,
-                                download,
-                              );
-                            } else {
-                              goToPostDetailsPage(
-                                context: context,
-                                posts: items,
-                                initialIndex: index,
-                                scrollController: _autoScrollController,
-                              );
-                            }
-                          }
-                        : null,
-                    isFaved: ref.watch(favoriteProvider(post.id)),
-                    enableFav: !multiSelect && canFavorite && !block,
-                    quickActionButtonBuilder:
-                        defaultImagePreviewButtonBuilder(ref, post),
-                    onFavToggle: (isFaved) async {
-                      if (isFaved) {
-                        if (favoriteAdder == null) return;
-                        await favoriteAdder(post.id, ref);
-                      } else {
-                        if (favoriteRemover == null) return;
-                        await favoriteRemover(post.id, ref);
-                      }
-                    },
-                    autoScrollOptions: AutoScrollOptions(
-                      controller: _autoScrollController,
-                      index: index,
-                    ),
-                    isAnimated: post.isAnimated,
-                    isTranslated: post.isTranslated,
-                    hasComments: post.hasComment,
-                    hasParentOrChildren: post.hasParentOrChildren,
-                    score: settings.showScoresInGrid ? post.score : null,
+                  autoScrollOptions: AutoScrollOptions(
+                    controller: _autoScrollController,
+                    index: index,
+                  ),
+                  isAnimated: post.isAnimated,
+                  isTranslated: post.isTranslated,
+                  hasComments: post.hasComment,
+                  hasParentOrChildren: post.hasParentOrChildren,
+                  score: settings.showScoresInGrid ? post.score : null,
+                  borderRadius: BorderRadius.circular(
+                    settings.imageBorderRadius,
+                  ),
+                  image: BooruImage(
+                    aspectRatio: post.aspectRatio,
+                    imageUrl: block
+                        ? ''
+                        : gridThumbnailUrlBuilder != null
+                            ? gridThumbnailUrlBuilder(settings, post)
+                            : post.thumbnailImageUrl,
                     borderRadius: BorderRadius.circular(
                       settings.imageBorderRadius,
                     ),
-                    image: BooruImage(
-                      aspectRatio: post.aspectRatio,
-                      imageUrl: block
-                          ? ''
-                          : gridThumbnailUrlBuilder != null
-                              ? gridThumbnailUrlBuilder(settings, post)
-                              : post.thumbnailImageUrl,
-                      borderRadius: BorderRadius.circular(
-                        settings.imageBorderRadius,
-                      ),
-                      forceFill:
-                          settings.imageListType == ImageListType.standard,
-                      placeholderUrl: post.thumbnailImageUrl,
-                      width: width,
-                      height: height,
-                      cacheHeight: cacheHeight,
-                      cacheWidth: cacheWidth,
-                    ),
+                    forceFill: settings.imageListType == ImageListType.standard,
+                    placeholderUrl: post.thumbnailImageUrl,
+                    width: width,
+                    height: height,
+                    cacheHeight: cacheHeight,
+                    cacheWidth: cacheWidth,
                   ),
                 ),
               ),
