+++ lib/core/scaffolds/post_details_page_scaffold.dart
@@ -118,150 +118,145 @@ class _PostDetailPageScaffoldState<T extends Post>
     final postGesturesHandler = booruBuilder?.postGestureHandlerBuilder;
 
     return LayoutBuilder(
-      builder: (context, constraints) => DownloadProviderWidget(
-        builder: (context, download) => DetailsPage(
-          sharedChildBuilder: (page) => Column(
-            children: [
-              if (widget.infoBuilder != null)
-                constraints.maxHeight > 450
-                    ? widget.infoBuilder!(context, posts[page])
-                    : const SizedBox.shrink(),
-              widget.toolbarBuilder != null
-                  ? widget.toolbarBuilder!(context, posts[page])
-                  : SimplePostActionToolbar(post: posts[page]),
-            ],
-          ),
-          controller: controller,
-          intitialIndex: widget.initialIndex,
-          onExit: widget.onExit,
-          onPageChanged: (page) {
-            onSwiped(page);
-            widget.onPageChangeIndexed?.call(page);
-            widget.onPageChanged?.call(posts[page]);
-          },
-          onSwipeDownEnd: booruBuilder?.canHandlePostGesture(
-                        GestureType.swipeDown,
-                        config.postGestures?.fullview,
-                      ) ==
-                      true &&
-                  postGesturesHandler != null
-              ? (page) => postGesturesHandler(
-                    ref,
-                    config.postGestures?.fullview?.swipeDown,
-                    posts[page],
-                    download,
-                  )
-              : null,
-          bottomSheet: (page, sharedChild) {
-            final bottomSheet = Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                if (posts[page].isVideo)
-                  ValueListenableBuilder(
-                    valueListenable: videoProgress,
-                    builder: (_, progress, __) => VideoSoundScope(
-                      builder: (context, soundOn) => BooruVideoProgressBar(
-                        soundOn: soundOn,
-                        progress: progress,
-                        playbackSpeed:
-                            ref.watchPlaybackSpeed(posts[page].videoUrl),
-                        onSeek: (position) => onVideoSeekTo(position, page),
-                        onSpeedChanged: (speed) =>
-                            ref.setPlaybackSpeed(posts[page].videoUrl, speed),
-                        onSoundToggle: (value) =>
-                            ref.setGlobalVideoSound(value),
-                      ),
-                    ),
-                  ),
-                if (sharedChild != null)
-                  Container(
-                    color: context.colorScheme.surface,
-                    padding: EdgeInsets.only(
-                      bottom: MediaQuery.paddingOf(context).bottom,
-                    ),
-                    child: sharedChild,
-                  ),
-              ],
-            );
-
-            return widget.infoBuilder != null
-                ? Container(
-                    decoration: BoxDecoration(
-                      color: context.theme.scaffoldBackgroundColor
-                          .withOpacity(0.8),
-                      border: Border(
-                        top: BorderSide(
-                          color: context.theme.dividerColor,
-                          width: 0.2,
-                        ),
-                      ),
-                    ),
-                    child: bottomSheet,
-                  )
-                : bottomSheet;
-          },
-          targetSwipeDownBuilder: (context, page) => SwipeTargetImage(
-            imageUrl: posts[page].isVideo
-                ? posts[page].videoThumbnailUrl
-                : widget.swipeImageUrlBuilder(posts[page]),
-            aspectRatio: posts[page].aspectRatio,
-          ),
-          expandedBuilder:
-              (context, page, currentPage, expanded, enableSwipe, sharedChild) {
-            final widgets = _buildWidgets(
-              context,
-              expanded,
-              page,
-              currentPage,
-              ref,
-              booruBuilder,
-              postGesturesHandler,
-              download,
-              sharedChild,
-            );
-
-            return Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 4),
-              child: CustomScrollView(
-                physics:
-                    enableSwipe ? null : const NeverScrollableScrollPhysics(),
-                controller: PageContentScrollController.of(context),
-                slivers: [
-                  SliverList(
-                    delegate: SliverChildBuilderDelegate(
-                      (context, index) => widgets[index],
-                      childCount: widgets.length,
-                    ),
-                  ),
-                  if (expanded && page == currentPage)
-                    if (widget.sliverArtistPostsBuilder != null)
-                      ...widget.sliverArtistPostsBuilder!(context, posts[page]),
-                  if (widget.sliverRelatedPostsBuilder != null &&
-                      ref.watch(_visibleProvider(currentPage)) &&
-                      expanded &&
-                      page == currentPage)
-                    widget.sliverRelatedPostsBuilder!(context, posts[page]),
-                  if (widget.sliverCharacterPostsBuilder != null &&
-                      expanded &&
-                      page == currentPage)
-                    widget.sliverCharacterPostsBuilder!(context, posts[page]),
-                  SliverSizedBox(
-                    height: MediaQuery.paddingOf(context).bottom + 72,
-                  ),
-                ],
-              ),
-            );
-          },
-          pageCount: widget.posts.length,
-          topRightButtonsBuilder: (page, expanded) =>
-              widget.topRightButtonsBuilder != null
-                  ? widget.topRightButtonsBuilder!(page, expanded, posts[page])
-                  : [
-                      GeneralMoreActionButton(post: widget.posts[page]),
-                    ],
-          onExpanded: (currentPage) =>
-              widget.onExpanded?.call(posts[currentPage]),
+      builder: (context, constraints) => DetailsPage(
+        sharedChildBuilder: (page) => Column(
+          children: [
+            if (widget.infoBuilder != null)
+              constraints.maxHeight > 450
+                  ? widget.infoBuilder!(context, posts[page])
+                  : const SizedBox.shrink(),
+            widget.toolbarBuilder != null
+                ? widget.toolbarBuilder!(context, posts[page])
+                : SimplePostActionToolbar(post: posts[page]),
+          ],
         ),
+        controller: controller,
+        intitialIndex: widget.initialIndex,
+        onExit: widget.onExit,
+        onPageChanged: (page) {
+          onSwiped(page);
+          widget.onPageChangeIndexed?.call(page);
+          widget.onPageChanged?.call(posts[page]);
+        },
+        onSwipeDownEnd: booruBuilder?.canHandlePostGesture(
+                      GestureType.swipeDown,
+                      config.postGestures?.fullview,
+                    ) ==
+                    true &&
+                postGesturesHandler != null
+            ? (page) => postGesturesHandler(
+                  ref,
+                  config.postGestures?.fullview?.swipeDown,
+                  posts[page],
+                )
+            : null,
+        bottomSheet: (page, sharedChild) {
+          final bottomSheet = Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              if (posts[page].isVideo)
+                ValueListenableBuilder(
+                  valueListenable: videoProgress,
+                  builder: (_, progress, __) => VideoSoundScope(
+                    builder: (context, soundOn) => BooruVideoProgressBar(
+                      soundOn: soundOn,
+                      progress: progress,
+                      playbackSpeed:
+                          ref.watchPlaybackSpeed(posts[page].videoUrl),
+                      onSeek: (position) => onVideoSeekTo(position, page),
+                      onSpeedChanged: (speed) =>
+                          ref.setPlaybackSpeed(posts[page].videoUrl, speed),
+                      onSoundToggle: (value) => ref.setGlobalVideoSound(value),
+                    ),
+                  ),
+                ),
+              if (sharedChild != null)
+                Container(
+                  color: context.colorScheme.surface,
+                  padding: EdgeInsets.only(
+                    bottom: MediaQuery.paddingOf(context).bottom,
+                  ),
+                  child: sharedChild,
+                ),
+            ],
+          );
+
+          return widget.infoBuilder != null
+              ? Container(
+                  decoration: BoxDecoration(
+                    color:
+                        context.theme.scaffoldBackgroundColor.withOpacity(0.8),
+                    border: Border(
+                      top: BorderSide(
+                        color: context.theme.dividerColor,
+                        width: 0.2,
+                      ),
+                    ),
+                  ),
+                  child: bottomSheet,
+                )
+              : bottomSheet;
+        },
+        targetSwipeDownBuilder: (context, page) => SwipeTargetImage(
+          imageUrl: posts[page].isVideo
+              ? posts[page].videoThumbnailUrl
+              : widget.swipeImageUrlBuilder(posts[page]),
+          aspectRatio: posts[page].aspectRatio,
+        ),
+        expandedBuilder:
+            (context, page, currentPage, expanded, enableSwipe, sharedChild) {
+          final widgets = _buildWidgets(
+            context,
+            expanded,
+            page,
+            currentPage,
+            ref,
+            booruBuilder,
+            postGesturesHandler,
+            sharedChild,
+          );
+
+          return Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 4),
+            child: CustomScrollView(
+              physics:
+                  enableSwipe ? null : const NeverScrollableScrollPhysics(),
+              controller: PageContentScrollController.of(context),
+              slivers: [
+                SliverList(
+                  delegate: SliverChildBuilderDelegate(
+                    (context, index) => widgets[index],
+                    childCount: widgets.length,
+                  ),
+                ),
+                if (expanded && page == currentPage)
+                  if (widget.sliverArtistPostsBuilder != null)
+                    ...widget.sliverArtistPostsBuilder!(context, posts[page]),
+                if (widget.sliverRelatedPostsBuilder != null &&
+                    ref.watch(_visibleProvider(currentPage)) &&
+                    expanded &&
+                    page == currentPage)
+                  widget.sliverRelatedPostsBuilder!(context, posts[page]),
+                if (widget.sliverCharacterPostsBuilder != null &&
+                    expanded &&
+                    page == currentPage)
+                  widget.sliverCharacterPostsBuilder!(context, posts[page]),
+                SliverSizedBox(
+                  height: MediaQuery.paddingOf(context).bottom + 72,
+                ),
+              ],
+            ),
+          );
+        },
+        pageCount: widget.posts.length,
+        topRightButtonsBuilder: (page, expanded) =>
+            widget.topRightButtonsBuilder != null
+                ? widget.topRightButtonsBuilder!(page, expanded, posts[page])
+                : [
+                    GeneralMoreActionButton(post: widget.posts[page]),
+                  ],
+        onExpanded: (currentPage) =>
+            widget.onExpanded?.call(posts[currentPage]),
       ),
     );
   }
@@ -274,7 +269,6 @@ class _PostDetailPageScaffoldState<T extends Post>
     WidgetRef ref,
     BooruBuilder? booruBuilder,
     PostGestureHandlerBuilder? postDetailsGesturesHandler,
-    DownloadDelegate download,
     Widget? sharedChild,
   ) {
     final post = posts[page];
@@ -298,7 +292,6 @@ class _PostDetailPageScaffoldState<T extends Post>
                 ref,
                 ref.watchConfig.postGestures?.fullview?.doubleTap,
                 post,
-                download,
               )
           : null,
       onLongPress: booruBuilder?.canHandlePostGesture(
@@ -311,7 +304,6 @@ class _PostDetailPageScaffoldState<T extends Post>
                 ref,
                 ref.watchConfig.postGestures?.fullview?.longPress,
                 post,
-                download,
               )
           : null,
       onCurrentVideoPositionChanged: onCurrentPositionChanged,
