+++ lib/core/widgets/downloads/download_post_button.dart
@@ -8,7 +8,6 @@ import 'package:material_symbols_icons/symbols.dart';
 // Project imports:
 import 'package:boorusama/core/feats/downloads/downloads.dart';
 import 'package:boorusama/core/feats/posts/posts.dart';
-import 'package:boorusama/core/widgets/widgets.dart';
 
 class DownloadPostButton extends ConsumerWidget {
   const DownloadPostButton({
@@ -22,33 +21,31 @@ class DownloadPostButton extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    return DownloadProviderWidget(
-      builder: (context, download) => !small
-          ? IconButton(
-              splashRadius: 16,
-              onPressed: () {
+    return !small
+        ? IconButton(
+            splashRadius: 16,
+            onPressed: () {
+              showDownloadStartToast(context);
+              ref.download(post);
+            },
+            icon: const Icon(
+              Symbols.download,
+            ),
+          )
+        : Material(
+            color: Colors.transparent,
+            child: InkWell(
+              onTap: () {
                 showDownloadStartToast(context);
-                download(post);
+                ref.download(post);
               },
-              icon: const Icon(
-                Symbols.download,
-              ),
-            )
-          : Material(
-              color: Colors.transparent,
-              child: InkWell(
-                onTap: () {
-                  showDownloadStartToast(context);
-                  download(post);
-                },
-                child: const Padding(
-                  padding: EdgeInsets.all(4),
-                  child: Icon(
-                    Symbols.download,
-                  ),
+              child: const Padding(
+                padding: EdgeInsets.all(4),
+                child: Icon(
+                  Symbols.download,
                 ),
               ),
             ),
-    );
+          );
   }
 }
