+++ lib/core/widgets/general_more_action_button.dart
@@ -10,7 +10,6 @@ import 'package:boorusama/core/feats/downloads/downloads.dart';
 import 'package:boorusama/core/feats/posts/posts.dart';
 import 'package:boorusama/core/router.dart';
 import 'package:boorusama/core/utils.dart';
-import 'package:boorusama/core/widgets/widgets.dart';
 import 'package:boorusama/foundation/i18n.dart';
 import 'package:boorusama/widgets/widgets.dart';
 
@@ -28,53 +27,50 @@ class GeneralMoreActionButton extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final booru = ref.watchConfig;
 
-    return DownloadProviderWidget(
-      builder: (context, download) => SizedBox(
-        width: 40,
-        child: Material(
-          color: Colors.black.withOpacity(0.5),
-          shape: const CircleBorder(),
-          child: BooruPopupMenuButton(
-            iconColor: Colors.white,
-            onSelected: (value) {
-              switch (value) {
-                case 'download':
-                  showDownloadStartToast(context);
-                  if (onDownload != null) {
-                    onDownload!(post);
-                  } else {
-                    download(post);
-                  }
-                  break;
-                case 'view_in_browser':
-                  launchExternalUrl(
-                    post.getUriLink(booru.url),
-                  );
-                  break;
-                case 'show_tag_list':
-                  goToShowTaglistPage(
-                    ref,
-                    post.extractTags(),
-                  );
-                  break;
-                case 'view_original':
-                  goToOriginalImagePage(context, post);
-                  break;
-                // ignore: no_default_cases
-                default:
-              }
-            },
-            itemBuilder: {
-              'download': const Text('download.download').tr(),
-              if (!booru.hasStrictSFW)
-                'view_in_browser':
-                    const Text('post.detail.view_in_browser').tr(),
-              'show_tag_list': const Text('View tags'),
-              if (post.hasFullView)
-                'view_original':
-                    const Text('post.image_fullview.view_original').tr(),
-            },
-          ),
+    return SizedBox(
+      width: 40,
+      child: Material(
+        color: Colors.black.withOpacity(0.5),
+        shape: const CircleBorder(),
+        child: BooruPopupMenuButton(
+          iconColor: Colors.white,
+          onSelected: (value) {
+            switch (value) {
+              case 'download':
+                showDownloadStartToast(context);
+                if (onDownload != null) {
+                  onDownload!(post);
+                } else {
+                  ref.download(post);
+                }
+                break;
+              case 'view_in_browser':
+                launchExternalUrl(
+                  post.getUriLink(booru.url),
+                );
+                break;
+              case 'show_tag_list':
+                goToShowTaglistPage(
+                  ref,
+                  post.extractTags(),
+                );
+                break;
+              case 'view_original':
+                goToOriginalImagePage(context, post);
+                break;
+              // ignore: no_default_cases
+              default:
+            }
+          },
+          itemBuilder: {
+            'download': const Text('download.download').tr(),
+            if (!booru.hasStrictSFW)
+              'view_in_browser': const Text('post.detail.view_in_browser').tr(),
+            'show_tag_list': const Text('View tags'),
+            if (post.hasFullView)
+              'view_original':
+                  const Text('post.image_fullview.view_original').tr(),
+          },
         ),
       ),
     );
