+++ lib/core/widgets/posts/default_multi_selection_actions.dart
@@ -2,15 +2,15 @@
 import 'package:flutter/material.dart';
 
 // Package imports:
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:material_symbols_icons/symbols.dart';
 
 // Project imports:
 import 'package:boorusama/core/feats/downloads/downloads.dart';
 import 'package:boorusama/core/feats/posts/posts.dart';
 import 'package:boorusama/core/pages/bookmarks/add_bookmarks_button.dart';
-import 'package:boorusama/core/widgets/widgets.dart';
 
-class DefaultMultiSelectionActions extends StatelessWidget {
+class DefaultMultiSelectionActions extends ConsumerWidget {
   const DefaultMultiSelectionActions({
     super.key,
     required this.selectedPosts,
@@ -21,25 +21,23 @@ class DefaultMultiSelectionActions extends StatelessWidget {
   final void Function() endMultiSelect;
 
   @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext context, WidgetRef ref) {
     return ButtonBar(
       alignment: MainAxisAlignment.center,
       children: [
-        DownloadProviderWidget(
-          builder: (context, download) => IconButton(
-            onPressed: selectedPosts.isNotEmpty
-                ? () {
-                    showDownloadStartToast(context);
-                    // ignore: prefer_foreach
-                    for (final p in selectedPosts) {
-                      download(p);
-                    }
-
-                    endMultiSelect();
+        IconButton(
+          onPressed: selectedPosts.isNotEmpty
+              ? () {
+                  showDownloadStartToast(context);
+                  // ignore: prefer_foreach
+                  for (final p in selectedPosts) {
+                    ref.download(p);
                   }
-                : null,
-            icon: const Icon(Symbols.download),
-          ),
+
+                  endMultiSelect();
+                }
+              : null,
+          icon: const Icon(Symbols.download),
         ),
         AddBookmarksButton(
           posts: selectedPosts,
