+++ lib/core/widgets/widgets.dart
@@ -20,7 +20,6 @@ export 'desktop_search_bar.dart';
 export 'details_page.dart';
 export 'details_page_desktop.dart';
 export 'downloads/download_post_button.dart';
-export 'downloads/download_provider_widget.dart';
 export 'error_box.dart';
 export 'favorite_tag_label_selector_field.dart';
 export 'general_more_action_button.dart';
